+++ Intune/Runbooks/changelog.md
@@ -2,6 +2,13 @@
 
 ## 2025
 
+### 08th of December
+
+* Update: Add-InfoToEntraDeviceExtensions.ps1 - Outsourced Autopatch exclude to its own script and removed any remarks about Autopatch
+* Add: Add-InfoToEntraDeviceExtensions.ps1 - "prevent overwrite" functionality for existing entries
+* Add: Add-ExcludeTagsToEntraDeviceExtensions.ps1 - New script to add exclude markers to exclude devices from Autopatch
+* New: Repair-PrimaryUser.ps1 - New script to repair primary user information on devices in Intune
+
 ### 25th of August
 
 * Add: Minor additions to the description